requirements.txt
Django==4.2.19
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
whitenoise==6.9.0
